feat(services): rename overseerr to seerr globally
