+++ internal/api/handlers/auth_nonce_test.go
@@ -46,4 +46,3 @@ func TestExtractJWTNonce(t *testing.T) {
 		}
 	})
 }
-
