+++ internal/services/manager/service_manager.go
@@ -13,8 +13,8 @@ import (
 	"github.com/autobrr/dashbrr/internal/models"
 	"github.com/autobrr/dashbrr/internal/services/cache"
 	"github.com/autobrr/dashbrr/internal/services/core"
-	"github.com/autobrr/dashbrr/internal/services/overseerr"
 	"github.com/autobrr/dashbrr/internal/services/plex"
+	"github.com/autobrr/dashbrr/internal/services/seerr"
 )
 
 // ServiceManager handles service initialization and data fetching
@@ -33,7 +33,7 @@ func NewServiceManager(db *database.DB, cache cache.Store) *ServiceManager {
 
 // InitializeService handles initial data fetching for a newly configured service
 func (m *ServiceManager) InitializeService(ctx context.Context, config *models.ServiceConfiguration) {
-	// Extract service type from instance ID (e.g., "overseerr-1" -> "overseerr")
+	// Extract service type from instance ID (e.g., "seerr-1" -> "seerr")
 	serviceType, ok := models.ServiceTypeFromInstanceID(config.InstanceID)
 	if !ok {
 		log.Debug().
@@ -52,8 +52,8 @@ func (m *ServiceManager) InitializeService(ctx context.Context, config *models.S
 	}
 
 	switch serviceType {
-	case "overseerr":
-		m.initializeOverseerr(ctx, config)
+	case "seerr":
+		m.initializeSeerr(ctx, config)
 	case "plex":
 		m.initializePlex(ctx, config)
 	// Add other service types here as needed
@@ -69,20 +69,20 @@ func (m *ServiceManager) InitializeService(ctx context.Context, config *models.S
 	}
 }
 
-// initializeOverseerr handles Overseerr-specific initialization
-func (m *ServiceManager) initializeOverseerr(ctx context.Context, config *models.ServiceConfiguration) {
+// initializeSeerr handles Seerr-specific initialization
+func (m *ServiceManager) initializeSeerr(ctx context.Context, config *models.ServiceConfiguration) {
 	// Check if we already have fresh data in cache
-	cacheKey := "overseerr:requests:" + config.InstanceID
+	cacheKey := "seerr:requests:" + config.InstanceID
 	var cachedData any
 	if err := m.cache.Get(ctx, cacheKey, &cachedData); err == nil {
 		log.Debug().
 			Str("instance", config.InstanceID).
-			Msg("Using cached Overseerr data")
+			Msg("Using cached Seerr data")
 		return
 	}
 
 	// Create service instance
-	service := &overseerr.OverseerrService{}
+	service := &seerr.SeerrService{}
 	service.SetDB(m.db)
 
 	// Fetch requests in a goroutine
@@ -96,7 +96,7 @@ func (m *ServiceManager) initializeOverseerr(ctx context.Context, config *models
 			log.Error().
 				Err(err).
 				Str("instance", config.InstanceID).
-				Msg("Failed to fetch initial Overseerr requests")
+				Msg("Failed to fetch initial Seerr requests")
 			return
 		}
 
@@ -105,13 +105,13 @@ func (m *ServiceManager) initializeOverseerr(ctx context.Context, config *models
 			log.Warn().
 				Err(err).
 				Str("instance", config.InstanceID).
-				Msg("Failed to cache Overseerr requests")
+				Msg("Failed to cache Seerr requests")
 			return
 		}
 
 		log.Debug().
 			Str("instance", config.InstanceID).
-			Msg("Successfully fetched and cached initial Overseerr requests")
+			Msg("Successfully fetched and cached initial Seerr requests")
 	}()
 }
