+++ internal/services/services.go
@@ -13,13 +13,13 @@ import (
 	_ "github.com/autobrr/dashbrr/internal/services/lidarr"
 	_ "github.com/autobrr/dashbrr/internal/services/maintainerr"
 	_ "github.com/autobrr/dashbrr/internal/services/nzbget"
-	_ "github.com/autobrr/dashbrr/internal/services/overseerr"
 	_ "github.com/autobrr/dashbrr/internal/services/plex"
 	_ "github.com/autobrr/dashbrr/internal/services/prowlarr"
 	_ "github.com/autobrr/dashbrr/internal/services/qui"
 	_ "github.com/autobrr/dashbrr/internal/services/radarr"
 	_ "github.com/autobrr/dashbrr/internal/services/readarr"
 	_ "github.com/autobrr/dashbrr/internal/services/sabnzbd"
+	_ "github.com/autobrr/dashbrr/internal/services/seerr"
 	_ "github.com/autobrr/dashbrr/internal/services/sonarr"
 	_ "github.com/autobrr/dashbrr/internal/services/tailscale"
 	_ "github.com/autobrr/dashbrr/internal/services/traefik"
